transformacoes - bug translacao inicial, bug rotacao

+++ ponto.cpp
@@ -1,3 +1,3 @@
 #include "ponto.h"
 
-Ponto::Ponto(int x, int y):x(x), y(y), coordW(1){}
+Ponto::Ponto(float x, float y):x(x), y(y), coordW(1){}
